+++ src/sprite_utils.c
@@ -6,7 +6,7 @@
 /*   By: levensta <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/29 22:42:05 by levensta          #+#    #+#             */
-/*   Updated: 2021/02/14 16:35:52 by levensta         ###   ########.fr       */
+/*   Updated: 2021/02/17 23:34:59 by levensta         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,8 +19,8 @@ static int	find_repeat_spr(t_all *cub, int x1, int y1)
 	i = 0;
 	while (i < cub->num_spr)
 	{
-		if ((int)floorf(cub->sprite[i].x) == x1 \
-		&& (int)floorf(cub->sprite[i].y) == y1)
+		if ((int)floor(cub->sprite[i].x) == x1 \
+		&& (int)floor(cub->sprite[i].y) == y1)
 			return (1);
 		i++;
 	}
@@ -47,23 +47,23 @@ void		find_sprite(t_all *cub, int x1, int y1)
 {
 	if (find_repeat_spr(cub, x1, y1))
 		return ;
-	cub->sprite[cub->num_spr].x = (float)x1 + 0.5f;
-	cub->sprite[cub->num_spr].y = (float)y1 + 0.5f;
+	cub->sprite[cub->num_spr].x = (double)x1 + 0.5f;
+	cub->sprite[cub->num_spr].y = (double)y1 + 0.5f;
 	cub->num_spr++;
 }
 
 void		find_dists(t_all *cub)
 {
 	int		i;
-	float	x1;
-	float	y1;
+	double	x1;
+	double	y1;
 
 	i = -1;
 	while (++i < cub->num_spr)
 	{
 		x1 = cub->sprite[i].x;
 		y1 = cub->sprite[i].y;
-		cub->sprite[i].distance = sqrtf(powf(x1 - cub->plr.x0, 2) + \
+		cub->sprite[i].distance = sqrt(powf(x1 - cub->plr.x0, 2) + \
 		powf(y1 - cub->plr.y0, 2));
 	}
 }
