+++ src/sprites.c
@@ -6,24 +6,24 @@
 /*   By: levensta <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/12 23:35:39 by levensta          #+#    #+#             */
-/*   Updated: 2021/02/14 18:10:34 by levensta         ###   ########.fr       */
+/*   Updated: 2021/02/17 23:32:37 by levensta         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "cub3d.h"
 
-static int	y_set(t_all *cub, int size, float *ty, t_sprite *d)
+static int	y_set(t_all *cub, int size, double *ty, t_sprite *d)
 {
 	int		y;
 
 	y = (cub->s_height - size) / 2;
-	y += (float)size / (float)cub->s_height;
+	y += (double)size / (double)cub->s_height;
 	*ty = 0.0f;
-	d->x = (float)cub->txt[4].width / (float)size;
-	d->y = (float)cub->txt[4].height / (float)size;
+	d->x = (double)cub->txt[4].width / (double)size;
+	d->y = (double)cub->txt[4].height / (double)size;
 	if (y < 0)
 	{
-		*ty = d->y * (float)(-y);
+		*ty = d->y * (double)(-y);
 		y = -1;
 	}
 	return (y);
@@ -32,7 +32,7 @@ static int	y_set(t_all *cub, int size, float *ty, t_sprite *d)
 void		draw_sprite(t_all *cub, t_sprite sp, int x_start, int x)
 {
 	t_sprite		d;
-	float			ty;
+	double			ty;
 	unsigned int	color;
 	int				y_start;
 	int				y;
@@ -47,7 +47,7 @@ void		draw_sprite(t_all *cub, t_sprite sp, int x_start, int x)
 		&& ty < cub->txt[4].height)
 		{
 			color = *(unsigned int*)(cub->txt[4].addr + ((int)ty * \
-			cub->txt[4].ll + (int)((float)x * d.x) * (cub->txt[4].bpp / 8)));
+			cub->txt[4].ll + (int)((double)x * d.x) * (cub->txt[4].bpp / 8)));
 			if (color != 0 && color != 0xFF000000 && \
 			cub->dists[x_start + x] >= sp.distance - 0.5f)
 				my_mlx_pixel_put(cub, x_start + x, y_start + y, color);
@@ -59,7 +59,7 @@ void		draw_sprite(t_all *cub, t_sprite sp, int x_start, int x)
 void		drawing_sprites(t_all *cub)
 {
 	t_sprite	dt;
-	float		ang;
+	double		ang;
 	int			tx;
 	int			i;
 
@@ -71,10 +71,10 @@ void		drawing_sprites(t_all *cub)
 	{
 		dt.y = cub->plr.y0 - cub->sprite[i].y;
 		dt.x = cub->plr.x0 - cub->sprite[i].x;
-		cub->sprite[i].angle = atan2f(dt.y, dt.x) - ang;
-		cub->sprite[i].size = cub->view_dist / (cosf(cub->sprite[i].angle) \
+		cub->sprite[i].angle = atan2(dt.y, dt.x) - ang;
+		cub->sprite[i].size = cub->view_dist / (cos(cub->sprite[i].angle) \
 		* cub->sprite[i].distance);
-		tx = cub->s_width / 2 + (tanf(cub->sprite[i].angle) * cub->view_dist) \
+		tx = cub->s_width / 2 + (tan(cub->sprite[i].angle) * cub->view_dist) \
 		- cub->sprite[i].size / 2;
 		if (cub->sprite[i].size <= 0 || tx >= cub->s_width)
 			continue ;
